+++ CHANGELOG.md
@@ -4,6 +4,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.5.4] / 2023-10-19
+### Features
+- Add `ComboBox` methods extension
+### Updated
+- Update `RibbonItemPanelExtension` and remove `Clone`.
+### ComboBox
+- Add SetWidth
+- Add SetCurrent
+- Add AddCurrentChanged
+- Add RemoveCurrentChanged
+- Add AddDropDownOpened
+- Add RemoveDropDownOpened
+- Add AddDropDownClosed
+- Add RemoveDropDownClosed
+### Fixed
+- Clone `RibbonItem` not clone events and break `TextBox`.
+### Tests
+- Add `RevitTextBoxTests`
+- Add `RevitComboBoxTests`
+
 ## [0.5.3] / 2023-07-13
 ## Updated
 - Update Image/LargeImage extension code.
@@ -255,6 +275,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[0.5.4]: ../../compare/0.5.3...0.5.4
 [0.5.3]: ../../compare/0.5.2...0.5.3
 [0.5.2]: ../../compare/0.5.1...0.5.2
 [0.5.1]: ../../compare/0.5.0...0.5.1